deps(deps): bump sigs.k8s.io/gateway-api from 1.3.0 to 1.4.1

+++ pkg/controller/mcpserver_controller.go
@@ -510,7 +510,7 @@ func (r *MCPReconciler) discoverServersFromHTTPRoutes(
 	if isExternal {
 		// use appProtocol from Service spec (standard k8s field)
 		for _, port := range service.Spec.Ports {
-			if backendRef.Port != nil && port.Port == int32(*backendRef.Port) {
+			if backendRef.Port != nil && port.Port == *backendRef.Port {
 				if port.AppProtocol != nil && strings.ToLower(*port.AppProtocol) == "https" {
 					protocol = "https"
 				}
